requirements/base.txt
redis==5.0.*
alembic==1.13.*
argon2-cffi==23.1.*
psycopg2==2.9.*
python-jose==3.3.*
SQLAlchemy[mypy]==2.0.20
httpx==0.27.*

pydantic[email]==2.7.*
pydantic-settings==2.2.*
psycopg[binary,pool]==3.1.*
fastapi==0.110.*
uvicorn[standard]==0.29.*

sentry-sdk==1.45.*
pandas==2.2.*
openpyxl==3.1.*
python-multipart==0.0.9
fastapi-sso==0.14.*